+++ src/Entity/ElementConstitutif.php
@@ -31,69 +31,84 @@ class ElementConstitutif
     #[ORM\Column]
     private ?int $id = null;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(length: 30, nullable: true, enumType: ModaliteEnseignementEnum::class)]
     private ?ModaliteEnseignementEnum $modaliteEnseignement = null;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(nullable: true)]
     private ?float $ects;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(nullable: true)]
     private ?float $volumeCmPresentiel;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(nullable: true)]
     private ?float $volumeTdPresentiel;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(nullable: true)]
     private ?float $volumeTpPresentiel;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(nullable: true)]
     private ?float $volumeCmDistanciel;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(nullable: true)]
     private ?float $volumeTdDistanciel;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(nullable: true)]
     private ?float $volumeTpDistanciel;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(nullable: true)]
     private ?bool $isCmPresentielMutualise;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(nullable: true)]
     private ?bool $isTdPresentielMutualise;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(nullable: true)]
     private ?bool $isTpPresentielMutualise;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(nullable: true)]
     private ?bool $isCmDistancielMutualise;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(nullable: true)]
     private ?bool $isTdDistancielMutualise;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(nullable: true)]
     private ?bool $isTpDistancielMutualise;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\ManyToOne]
     private ?NatureUeEc $natureUeEc = null;
 
     #[ORM\OneToMany(mappedBy: 'ec', targetEntity: Mccc::class, cascade: ['persist', 'remove'], orphanRemoval: true)]
     private Collection $mcccs;
 
-    #[Groups(['DTO_json_versioning'])]
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(length: 15)]
     private ?string $code = null;
 
-    #[Groups(['DTO_json_versioning'])]
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column]
     private ?int $ordre = null;
 
-    #[Groups(['DTO_json_versioning'])]
+    #[Groups('parcours_json_versioning')]
     #[ORM\ManyToOne(fetch: 'EAGER', inversedBy: 'elementConstitutifs')]
     private ?FicheMatiere $ficheMatiere = null;
 
     #[ORM\ManyToOne(inversedBy: 'elementConstitutifs')]
     private ?Parcours $parcours = null;
-  
+
     #[ORM\ManyToOne(inversedBy: 'elementConstitutifs', fetch: 'EAGER')]
     private ?Ue $ue = null;
 
@@ -101,7 +116,7 @@ class ElementConstitutif
     #[ORM\Column(length: 255, nullable: true)]
     private ?string $texteEcLibre = null;
 
-    #[Groups(['DTO_json_versioning'])]
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(length: 255, nullable: true)]
     private ?string $libelle = null;
 
@@ -110,12 +125,12 @@ class ElementConstitutif
     private ?TypeEc $typeEc = null;
 
     #[MaxDepth(1)]
-    #[Groups(['DTO_json_versioning'])]
+    #[Groups('parcours_json_versioning')]
     #[ORM\ManyToOne(targetEntity: self::class, inversedBy: 'ecEnfants')]
     private ?self $ecParent = null;
 
     #[MaxDepth(1)]
-    #[Groups(['DTO_json_versioning'])]
+    #[Groups('parcours_json_versioning')]
     #[ORM\OneToMany(mappedBy: 'ecParent', targetEntity: self::class, cascade: ['persist', 'remove'], orphanRemoval: true)]
     #[ORM\OrderBy(['ordre' => 'ASC'])]
     private Collection $ecEnfants;
@@ -126,12 +141,14 @@ class ElementConstitutif
     #[ORM\Column(length: 255, nullable: true)]
     private ?string $etatMccc = null;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(nullable: true)]
     private ?float $volumeTe = null;
 
     #[ORM\Column]
     private ?bool $mcccEnfantsIdentique = false;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column]
     private ?bool $heuresEnfantsIdentiques = false;
 
@@ -148,17 +165,32 @@ class ElementConstitutif
     #[ORM\Column(nullable: true)]
     private ?bool $synchroMccc = null;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(nullable: true)]
     private ?bool $synchroHeures = null;
 
     #[ORM\Column(nullable: true)]
     private ?bool $synchroBcc = null;
-    
+
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(nullable: true)]
     private ?bool $synchroEcts = null;
 
+    #[Groups('parcours_json_versioning')]
     #[ORM\Column(nullable: true)]
     private ?bool $sansHeure = false;
+
+    /**
+     * VERSIONING
+     *
+     * Ceci n'est pas un enregistrement BD
+     */
+    #[Groups('parcours_json_versioning')]
+    public ?bool $raccroche = null;
+
+    #[ORM\Column(length: 8, nullable: true)]
+    private ?string $codeApogee = null;
+
 
     public function __construct()
     {
@@ -804,5 +836,37 @@ class ElementConstitutif
         $this->sansHeure = $sansHeure;
 
         return $this;
+    }
+
+    public function isRaccroche() : bool {
+        if($this->parcours){
+            $this->raccroche = GetElementConstitutif::isRaccroche($this, $this->parcours);
+        }
+        return $this->raccroche;
+    }
+
+    public function getCodeApogee(): ?string
+    {
+        return $this->codeApogee;
+    }
+
+    public function setCodeApogee(?string $codeApogee): static
+    {
+        $this->codeApogee = $codeApogee;
+
+        return $this;
+    }
+
+    public function displayCodeApogee(): string
+    {
+        if ($this->codeApogee !== null) {
+            return $this->codeApogee ?? 'Aucun code Apogée';
+        }
+
+        if ($this->ficheMatiere !== null) {
+            return $this->ficheMatiere->getCodeApogee() ?? 'Aucun code Apogée';
+        }
+
+        return 'Aucun code Apogée';
     }
 }