+++ src/Enums/Apogee/CodeNatuElpEnum.php
@@ -1,0 +1,22 @@
+<?php
+/*
+ * Copyright (c) 2024. | David Annebicque | ORéOF  - All Rights Reserved
+ * @file /Users/davidannebicque/Sites/oreof/src/Enums/Apogee/CodeNatuElpEnum.php
+ * @author davidannebicque
+ * @project oreof
+ * @lastUpdate 16/01/2024 19:30
+ */
+
+namespace App\Enums\Apogee;
+
+enum CodeNatuElpEnum: string
+{
+    //SEM, UE, CHOI, MATI, MATM, MATS, MATP
+    case SEM = 'SEM';
+    case UE = 'UE';
+    case CHOI = 'CHOI';
+    case MATI = 'MATI';
+    case MATM = 'MATM';
+    case MATS = 'MATS';
+    case MATP = 'MATP';
+}